The commands here are a first attempt to streamline Cozmo setup on OS X for me and other people that have run into trouble installing Cozmo on OS X. Read the bash scripts first, before you decide to run them. You can also copy the lines you want to run.
From a clean install of OS X you can follow steps 1 to 4. I highly
recommend running 2b first on systems that already have HomeBrew.
Ideally you try stuff like this in a virtual machine with a copy of
your setup.

Installing Cozmo on OSX can be challenging, especially when different
version of Python IDE’s are installed. In 2a there is a reference to
and in 5c this command is run: brew reinstall python3 --with-tcl-tk.
This will reinstall Python 3 with tkinter.

On different Mac’s I removed the Anaconda environment first to get
Cozmo to play nice with my setup. A bit of a rigorous move, so you have
to decide for yourself if you want to do that. You can try to add it to
your system again once you have Cozmo running (if you need it). People
more versed in dealing with symlink trouble can find alternative
solutions, I am open to suggestions.